+++ src/main.ts
@@ -129,7 +129,7 @@ function renderTitle() {
     </div>`);
 }
 
-const VERSION = "0.11.0";
+const VERSION = "0.12.0";
 
 function renderHowTo() {
   openModal("How to Play", `